mirror: make mirror legs on different tag pvs
The mirror legs could be placed on the PVs(with different
tags) from the other mirror legs with this patch.

The lvcreate allocates the replica "randomly" when creating
a mirrored logical volume now, so it could happen that both
copies end up on the same array (or on the same site), clearly
undesired.

We can't control the fist allocation, because didn't restrict
the areas into parallel space. This patch add the limit when
finding areas for parallel space, so that mirror legs will
always be placed on pvs with differnt tags.

Signed-off-by: Guangliang Zhao <gzhao@suse.com>