clvmd: don't report error on testing connection
In HA cluster, we have "clvm" resource agent to manage clvmd daemon.
The agent invokes clvmd like: "clvmd -T90 -d0", which  always prints
a scaring error message:

"""
local socket: connect failed: No such file or directory
"""

When specifed with "-d" option, clvmd tries to check if an instance
of the clvmd daemon is already running through a testing connection.
The connect() will fail with this error, which is expected.

I don't see the necessary to print error on testing connection, so
remove the error.

Signed-off-by: Eric Ren <zren@suse.com>